MMI rename folder
